test: Player에 있던 Gamer 테스트 옮김
